+++ src/main/java/com/github/merchantpug/toomanyorigins/TooManyOrigins.java
@@ -24,14 +24,17 @@ SOFTWARE.
 
 package com.github.merchantpug.toomanyorigins;
 
+import com.github.merchantpug.apugli.Apugli;
+import com.github.merchantpug.apugli.util.ApugliConfig;
+import com.github.merchantpug.apugli.util.ApugliServerConfig;
+import com.github.merchantpug.toomanyorigins.integration.LegacyContentRegistrationEvent;
 import com.github.merchantpug.toomanyorigins.networking.TMOPacketsC2S;
 import com.github.merchantpug.toomanyorigins.registry.*;
 import com.github.merchantpug.toomanyorigins.util.TooManyOriginsServerConfig;
+import eu.midnightdust.lib.config.MidnightConfig;
 import io.github.apace100.apoli.util.NamespaceAlias;
 import com.github.merchantpug.toomanyorigins.util.TooManyOriginsConfig;
-import com.github.merchantpug.apugli.Apugli;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
+import io.github.apace100.origins.integration.OriginDataLoadedCallback;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.util.Identifier;
@@ -40,12 +43,13 @@ import org.apache.logging.log4j.Logger;
 
 public class TooManyOrigins implements ModInitializer {
 	public static final String MODID = "toomanyorigins";
-	public static final Logger LOGGER = LogManager.getLogger(TooManyOrigins.class);
+	public static final Logger LOGGER = LogManager.getLogger("TooManyOrigins");
 	public static String VERSION = "";
 	public static int[] SEMVER;
 
-	public static TooManyOriginsConfig config;
-	public static TooManyOriginsServerConfig serverConfig;
+	public static boolean legacyDragonbornContentRegistered = false;
+	public static boolean legacyUndeadContentRegistered = false;
+	public static boolean legacyWitheredContentRegistered = false;
 
 	@Override
 	public void onInitialize() {
@@ -64,6 +68,9 @@ public class TooManyOrigins implements ModInitializer {
 			}
 		});
 		LOGGER.info("TooManyOrigins " + VERSION + " is initializing. Enjoy!");
+
+		MidnightConfig.init(TooManyOrigins.MODID, TooManyOriginsConfig.class);
+
 		TMOBlocks.register();
 		TMOEffects.register();
 		TMOEntities.register();
@@ -71,16 +78,22 @@ public class TooManyOrigins implements ModInitializer {
 		TMOPowers.register();
 		TMOSounds.register();
 
-		Apugli.addDependent(MODID, "1.7.2");
 		NamespaceAlias.addAlias(MODID, "apugli");
 
-		AutoConfig.register(TooManyOriginsConfig.class, JanksonConfigSerializer::new);
-		config = AutoConfig.getConfigHolder(TooManyOriginsConfig.class).getConfig();
-
-		AutoConfig.register(TooManyOriginsServerConfig.class, JanksonConfigSerializer::new);
-		serverConfig = AutoConfig.getConfigHolder(TooManyOriginsServerConfig.class).getConfig();
-
 		TMOPacketsC2S.register();
+		OriginDataLoadedCallback.EVENT.register(LegacyContentRegistrationEvent::onDataLoaded);
+
+		if (TooManyOriginsConfig.legacyDragonbornEnabled) {
+			legacyDragonbornContentRegistered = true;
+		}
+
+		if (TooManyOriginsConfig.legacyUndeadEnabled) {
+			legacyUndeadContentRegistered = true;
+		}
+
+		if (TooManyOriginsConfig.legacyWitheredEnabled) {
+			legacyWitheredContentRegistered = true;
+		}
 	}
 
 	public static Identifier identifier(String path) {