+++ src/main/java/com/github/merchantpug/toomanyorigins/blocks/WitheredStemBlock.java
@@ -6,11 +6,11 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.random.Random;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
 
-import java.util.Random;
 import java.util.function.Supplier;
 
 public class WitheredStemBlock extends PlantBlock {