+++ src/main/java/com/github/merchantpug/toomanyorigins/mixin/SittingFlamingPhaseMixin.java
@@ -2,6 +2,7 @@ package com.github.merchantpug.toomanyorigins.mixin;
 
 import com.github.merchantpug.toomanyorigins.registry.TMOEffects;
 import com.github.merchantpug.toomanyorigins.TooManyOrigins;
+import com.github.merchantpug.toomanyorigins.util.TooManyOriginsConfig;
 import net.minecraft.entity.boss.dragon.phase.SittingFlamingPhase;
 import net.minecraft.entity.effect.StatusEffect;
 import org.spongepowered.asm.mixin.Mixin;
@@ -12,7 +13,7 @@ import org.spongepowered.asm.mixin.injection.ModifyArg;
 public class SittingFlamingPhaseMixin {
     @ModifyArg(method = "serverTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/effect/StatusEffectInstance;<init>(Lnet/minecraft/entity/effect/StatusEffect;)V"))
     private StatusEffect statusEffect(StatusEffect effect) {
-        if (TooManyOrigins.config.dragonFireballMixin.shouldFireballDamageUndead) {
+        if (TooManyOriginsConfig.shouldFireballDamageUndead) {
             return effect = TMOEffects.END_FIRE;
         }
         return effect;