+++ src/main/java/com/github/merchantpug/toomanyorigins/registry/TMOItems.java
@@ -1,6 +1,7 @@
 package com.github.merchantpug.toomanyorigins.registry;
 
 import com.github.merchantpug.toomanyorigins.items.DragonFireballItem;
+import com.github.merchantpug.toomanyorigins.util.TooManyOriginsConfig;
 import net.minecraft.item.AliasedBlockItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroup;
@@ -11,20 +12,26 @@ import com.github.merchantpug.toomanyorigins.TooManyOrigins;
 
 public class TMOItems {
 
-    public static final Item DRAGON_FIREBALL = registerItem("dragon_fireball",
-            new DragonFireballItem((new Item.Settings()).maxCount(16).group(ItemGroup.COMBAT)));
+    public static Item DRAGON_FIREBALL;
 
-    public static final Item WITHERED_CROP_SEEDS = registerItem("withered_crop_seeds",
-            new AliasedBlockItem(TMOBlocks.WITHERED_CROP, (new Item.Settings()).group(ItemGroup.MATERIALS)));
+    public static Item WITHERED_CROP_SEEDS;
 
-    public static final Item WITHERED_STEM_SEEDS = registerItem("withered_stem_seeds",
-            new AliasedBlockItem(TMOBlocks.WITHERED_STEM, (new Item.Settings()).group(ItemGroup.MATERIALS)));
+    public static Item WITHERED_STEM_SEEDS;
 
 
     private static Item registerItem(String name, Item item) {
         return Registry.register(Registry.ITEM, new Identifier(TooManyOrigins.MODID, name), item);
     }
     public static void register() {
-
+        if (TooManyOriginsConfig.legacyDragonbornEnabled) {
+            DRAGON_FIREBALL = registerItem("dragon_fireball",
+                    new DragonFireballItem((new Item.Settings()).maxCount(16).group(ItemGroup.COMBAT)));
+        }
+        if (TooManyOriginsConfig.legacyWitheredEnabled) {
+           WITHERED_CROP_SEEDS = registerItem("withered_crop_seeds",
+                    new AliasedBlockItem(TMOBlocks.WITHERED_CROP, (new Item.Settings()).group(ItemGroup.MATERIALS)));
+           WITHERED_STEM_SEEDS = registerItem("withered_stem_seeds",
+                    new AliasedBlockItem(TMOBlocks.WITHERED_STEM, (new Item.Settings()).group(ItemGroup.MATERIALS)));
+        }
     }
 }