+++ src/main/java/com/github/merchantpug/toomanyorigins/util/TooManyOriginsServerConfig.java
@@ -1,12 +1,8 @@
 package com.github.merchantpug.toomanyorigins.util;
 
-import com.github.merchantpug.apugli.Apugli;
-import com.github.merchantpug.toomanyorigins.TooManyOrigins;
-import me.shedaniel.autoconfig.ConfigData;
-import me.shedaniel.autoconfig.annotation.Config;
-import me.shedaniel.autoconfig.annotation.ConfigEntry;
+import eu.midnightdust.lib.config.MidnightConfig;
 
-@Config(name = TooManyOrigins.MODID + "_server")
-public class TooManyOriginsServerConfig implements ConfigData {
-    public boolean performVersionCheck = true;
+public class TooManyOriginsServerConfig extends MidnightConfig {
+    @Entry
+    public static boolean performVersionCheck = true;
 }